reduce img res to 250x250 & remove colours
I used GIMP to resize the image down to 250x250 from 500x500. After
that, I removed all of the colour from the image, because it looked like
the colour was clashing with the new dark theme.

I'd say it looks a bit better now? The main goal here was to reduce the
image size and it's definitely gone down now. Something like 200KB down
to 12KB - I'd say that's worth losing a bit of quality in the pic?